Factor out swifttoolchain module and use in microwendy
