+++ go/internal/cli/commands/docker_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"net"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -456,108 +455,4 @@ func TestFindIPv4ViaNeighborTable_UnknownAddress(t *testing.T) {
 	// commands and read the host's neighbor tables, making it environment-dependent.
 	// Skip it to avoid flakiness/timeouts in unit test environments.
 	t.Skip("disabled: findIPv4ViaNeighborTable depends on host neighbor tables and OS commands")
-}
-
-func TestEnsureSwiftVersion_AlreadyInstalled(t *testing.T) {
-	original := execCommandContext
-	t.Cleanup(func() { execCommandContext = original })
-
-	var calls [][]string
-	execCommandContext = func(ctx context.Context, name string, args ...string) *exec.Cmd {
-		calls = append(calls, append([]string{name}, args...))
-		return exec.CommandContext(ctx, "true")
-	}
-
-	if err := ensureSwiftVersion(context.Background()); err != nil {
-		t.Fatalf("ensureSwiftVersion() unexpected error: %v", err)
-	}
-
-	// When "swiftly which" succeeds, no install should happen.
-	if len(calls) != 1 {
-		t.Fatalf("expected 1 call (which), got %d: %v", len(calls), calls)
-	}
-	if calls[0][0] != "swiftly" || calls[0][1] != "which" || calls[0][2] != defaultSwiftVersion {
-		t.Errorf("expected [swiftly which %s], got %v", defaultSwiftVersion, calls[0])
-	}
-}
-
-func TestEnsureSwiftVersion_InstallsWhenMissing(t *testing.T) {
-	original := execCommandContext
-	t.Cleanup(func() { execCommandContext = original })
-
-	var calls [][]string
-	execCommandContext = func(ctx context.Context, name string, args ...string) *exec.Cmd {
-		call := append([]string{name}, args...)
-		calls = append(calls, call)
-		// "swiftly which" fails (version not installed), "swiftly install" succeeds.
-		if len(args) > 0 && args[0] == "which" {
-			return exec.CommandContext(ctx, "false")
-		}
-		return exec.CommandContext(ctx, "true")
-	}
-
-	if err := ensureSwiftVersion(context.Background()); err != nil {
-		t.Fatalf("ensureSwiftVersion() unexpected error: %v", err)
-	}
-
-	if len(calls) != 2 {
-		t.Fatalf("expected 2 calls (which + install), got %d: %v", len(calls), calls)
-	}
-	if calls[0][1] != "which" {
-		t.Errorf("first call should be which, got %v", calls[0])
-	}
-	if calls[1][1] != "install" || calls[1][2] != defaultSwiftVersion {
-		t.Errorf("expected [swiftly install %s], got %v", defaultSwiftVersion, calls[1])
-	}
-}
-
-func TestEnsureSwiftVersion_SwiftlyNotFound(t *testing.T) {
-	original := execCommandContext
-	t.Cleanup(func() { execCommandContext = original })
-
-	execCommandContext = func(ctx context.Context, name string, args ...string) *exec.Cmd {
-		return exec.CommandContext(ctx, "nonexistent-binary-that-does-not-exist")
-	}
-
-	err := ensureSwiftVersion(context.Background())
-	if err == nil {
-		t.Fatal("ensureSwiftVersion() expected error when swiftly not found, got nil")
-	}
-	if !strings.Contains(err.Error(), "swiftly is required but not installed") {
-		t.Errorf("expected actionable error message, got: %v", err)
-	}
-}
-
-func TestEnsureSwiftVersion_InstallFails(t *testing.T) {
-	original := execCommandContext
-	t.Cleanup(func() { execCommandContext = original })
-
-	execCommandContext = func(ctx context.Context, name string, args ...string) *exec.Cmd {
-		return exec.CommandContext(ctx, "false")
-	}
-
-	err := ensureSwiftVersion(context.Background())
-	if err == nil {
-		t.Fatal("ensureSwiftVersion() expected error on install failure, got nil")
-	}
-	if !strings.Contains(err.Error(), "installing Swift") {
-		t.Errorf("expected install error message, got: %v", err)
-	}
-}
-
-func TestEnsureSwiftVersion_Cancellation(t *testing.T) {
-	original := execCommandContext
-	t.Cleanup(func() { execCommandContext = original })
-
-	execCommandContext = func(ctx context.Context, name string, args ...string) *exec.Cmd {
-		return exec.CommandContext(ctx, "true")
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	err := ensureSwiftVersion(ctx)
-	if err == nil {
-		t.Fatal("ensureSwiftVersion() expected error on cancelled context, got nil")
-	}
 }
